Initial commit: Launching ServerlessLLM
This release includes basic modules of ServerlessLLM.

This project builds upon our research published in OSDI 2024.
As we transitioned from an internal development environment to this open-source platform, the detailed commit history and authorship information could not be preserved. Hence, we are taking this opportunity to formally acknowledge all contributors.

 - [Yao Fu](https://github.com/future-xy)
 - [Leyang Xue](https://github.com/drunkcoding)
 - [Yeqi Huang](https://github.com/Chivier)
 - [Andrei-Octavian Brabete](https://github.com/andrei3131)
 - [Matej Sandor](https://github.com/MatejSandor)
 - [Ruiqi Lai](https://github.com/lrq619)
 - [Siyang Shao](https://github.com/SiyangShao)
 - [Xinyuan Tong](https://github.com/JustinTong0323)

Thank you to everyone involved for your dedication and expertise.