[Feat] Enhance deploy command to support partial configurations (#46)
### Description

This PR enhances the deploy functionality in `sllm-cli` to support
providing partial configurations. Missing sections in the user-provided
configuration will be filled in using the default configuration, making
the deployment process more flexible and testing easier.

### Changes

- Updated `deploy.py` to: Fully read and merge the provided incomplete
configuration with the default configuration.

### Example Usage

Users can now deploy a model with partial configurations. For instance:

```json
{
    "model": "facebook/opt-1.3b",
    "backend": "transformers",
    "num_gpus": 1,
    "backend_config": {
        "pretrained_model_name_or_path": "facebook/opt-1.3b",
        "device_map": "auto",
        "torch_dtype": "float16"
    }
}
```

In this example, the provided configuration does not include
`auto_scaling_config`, so the default configuration for
`auto_scaling_config` will be used.

This PR is related to #23  case 2.

---------

Signed-off-by: Xinyuan Tong <justinning0323@outlook.com>